Force int for loan_id

+++ loan-del.php
@@ -10,7 +10,7 @@ require_once('module/html-functions.php');
 auth_or_login('loan-list.php');
 level_or_alert(3, 'Suppression d’un prêt');
 
-$loan_id = param_post_or_get('loan_id', 0);
+$loan_id = (int)param_post_or_get('loan_id', 0);
 $valid   = param_post('ok', 'no');
 
 if ($loan_id == 0 || $valid == 'cancel')
